+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable remote write receiver.
 - Add prometheus remote write ingress boilerplate.
 - Add prometheus agent remote write secret boilerplate.
+- Generate prometheus remote write agent secret and config.
 
 ## [4.6.2] - 2022-09-13